requirements.txt
face_recognition==1.3.0
opencv-python==4.6.0.66
pillow==9.2.0
mediapipe